+++ CMakeLists.txt
@@ -2749,6 +2749,7 @@ foreach(_hdr
   include/grpcpp/security/tls_credentials_options.h
   include/grpcpp/server.h
   include/grpcpp/server_builder.h
+  include/grpcpp/server_builder_impl.h
   include/grpcpp/server_context.h
   include/grpcpp/server_impl.h
   include/grpcpp/server_posix.h
@@ -3441,6 +3442,7 @@ foreach(_hdr
   include/grpcpp/security/tls_credentials_options.h
   include/grpcpp/server.h
   include/grpcpp/server_builder.h
+  include/grpcpp/server_builder_impl.h
   include/grpcpp/server_context.h
   include/grpcpp/server_impl.h
   include/grpcpp/server_posix.h
