+++ Makefile
@@ -4946,6 +4946,7 @@ PUBLIC_HEADERS_CXX += \
     include/grpcpp/security/tls_credentials_options.h \
     include/grpcpp/server.h \
     include/grpcpp/server_builder.h \
+    include/grpcpp/server_builder_impl.h \
     include/grpcpp/server_context.h \
     include/grpcpp/server_impl.h \
     include/grpcpp/server_posix.h \
@@ -5643,6 +5644,7 @@ PUBLIC_HEADERS_CXX += \
     include/grpcpp/security/tls_credentials_options.h \
     include/grpcpp/server.h \
     include/grpcpp/server_builder.h \
+    include/grpcpp/server_builder_impl.h \
     include/grpcpp/server_context.h \
     include/grpcpp/server_impl.h \
     include/grpcpp/server_posix.h \
