+++ include/grpcpp/impl/codegen/completion_queue_impl.h
@@ -48,6 +48,7 @@ namespace grpc_impl {
 
 class Channel;
 class Server;
+class ServerBuilder;
 template <class R>
 class ClientReader;
 template <class W>
@@ -78,7 +79,6 @@ class ErrorMethodHandler;
 namespace grpc {
 
 class ChannelInterface;
-class ServerBuilder;
 class ServerInterface;
 
 namespace internal {
@@ -255,7 +255,7 @@ class CompletionQueue : private ::grpc::GrpcLibraryCodegen {
  private:
   // Friends for access to server registration lists that enable checking and
   // logging on shutdown
-  friend class ::grpc::ServerBuilder;
+  friend class ::grpc_impl::ServerBuilder;
   friend class ::grpc_impl::Server;
 
   // Friend synchronous wrappers so that they can access Pluck(), which is
@@ -445,7 +445,7 @@ class ServerCompletionQueue : public CompletionQueue {
         polling_type_(polling_type) {}
 
   grpc_cq_polling_type polling_type_;
-  friend class ::grpc::ServerBuilder;
+  friend class ::grpc_impl::ServerBuilder;
   friend class ::grpc_impl::Server;
 };
