+++ tools/doxygen/Doxyfile.c++
@@ -1033,6 +1033,7 @@ include/grpcpp/security/server_credentials_impl.h \
 include/grpcpp/security/tls_credentials_options.h \
 include/grpcpp/server.h \
 include/grpcpp/server_builder.h \
+include/grpcpp/server_builder_impl.h \
 include/grpcpp/server_context.h \
 include/grpcpp/server_impl.h \
 include/grpcpp/server_posix.h \
